Move queue to `lifo` module, update tests+benches
